+++ exporter.py
@@ -13,7 +13,7 @@ from exporter.queue.config import AmqpConnConfig, QueueConfig
 from exporter.schema import SchemaService
 from exporter.session_context import configure_logger
 from exporter.terra.client import TerraClient
-from exporter.terra.terra_exporter import TerraExporter
+from exporter.terra.exporter import TerraExporter
 from exporter.terra.terra_listener import TerraListener
 from manifest.exporter import ManifestExporter
 from manifest.generator import ManifestGenerator
