+++ exporter/terra/terra_listener.py
@@ -11,7 +11,7 @@ from exporter.ingest.service import IngestService
 from exporter.queue.config import QueueConfig, AmqpConnConfig
 from exporter.session_context import SessionContext
 from exporter.terra.exceptions import ExperimentMessageParseException
-from exporter.terra.terra_exporter import TerraExporter
+from exporter.terra.exporter import TerraExporter
 
 
 @dataclass
