+++ tests/exporter/terra/test_terra_listener.py
@@ -6,7 +6,7 @@ from mock import MagicMock
 
 from exporter.ingest.service import IngestService
 from exporter.queue.config import QueueConfig
-from exporter.terra.terra_exporter import TerraExporter
+from exporter.terra.exporter import TerraExporter
 from exporter.terra.terra_listener import _TerraListener
